+++ spec/identifiers/isbn_spec.rb
@@ -121,6 +121,14 @@ RSpec.describe Identifiers::ISBN do
     expect(described_class.extract('0-80506909-7')).to be_empty
   end
 
+  it 'does not extract ISBNs from a long run of digits separated by hyphens' do
+    expect(described_class.extract('0-1884-0-3140-0-4396-0-5652-0-4396-0-2826')).to be_empty
+  end
+
+  it 'does not extract ISBNs joined to other digits by a hyphen' do
+    expect(described_class.extract('9780805069099-9780671879198')).to be_empty
+  end
+
   it 'extracts ISBN-10s with variable width registration group identifiers' do
     expect(described_class.extract('99921-58-10-7 9971-5-0210-0 960-425-059-0 80-902734-1-6'))
       .to contain_exactly('9789992158104', '9789971502102', '9789604250592', '9788090273412')